secretbox encryption of opaque objects..
so that now they're actually opaque. Secret key is provided to client as
a third-party caveat condition addressed to the client.

There are some nuances to this that are interesting. For example, we
could obtain a macaroon without this caveat, for clients that want to be
able to delete the object, but should not have access to decrypt it.

This changeset requires a slight modification to expose
httpbakery.ObtainThirdPartyCaveat, so it can't land yet.